Other targets: upgrade vararg matching for new FormatString
Enhance FormatString declaration to validate arg/convspec correspondence when __GNUC__. In main target and elsewhere, make necessary changes mandated by diagnostics so issued.

There's a lot of stuff in TLEdit that is deprecated in latest runtimes, so TLEdit continues in 10.15 (Catalina).

+++ TLEdit/GenericWindlgController.mm
@@ -294,7 +294,7 @@ static LoadFiascoProtectedUnorderedMap<int, dlgInitData> InitDataMap;
 -(HWND)GetControlHWND:(NSInteger)ctlid
 {
     if (CtlidToHWND.count(ctlid) == 0)
-        [self barf: FormatString("Received action from Cocoa control tagged %d, for which no emulation exists.", ctlid)];
+        [self barf: FormatString("Received action from Cocoa control tagged %d, for which no emulation exists.", (int)ctlid)];
     return CtlidToHWND[ctlid];
 }
 -(NSView*)GetControlView:(NSInteger)ctlid
